\name{robustgam-package}
\alias{robustgam-package}
\docType{package}
\title{
Robust Estimation for Generalized Additive Models
}
\description{
This package provides robust estimation for generalized additive models. It implements a fast and stable algorithm in Wong, Yao and Lee (2012). The implementation also contains three automatic selection methods for smoothing parameter. They are designed to be robust to outliers. For more details, see Wong, Yao and Lee (2012).
}
\details{
\tabular{ll}{
Package: \tab robustgam\cr
Type: \tab Package\cr
Version: \tab 0.1.4\cr
Date: \tab 2012-11-17\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Raymond K. W. Wong
Maintainer: Raymond K. W. Wong <raymondkww.dev@gmail.com>
}
\references{
Raymond K. W. Wong, Fang Yao and Thomas C. M. Lee (2012) \emph{Robust Estimation for Generalized Additive Models}. Submitted to \emph{Journal of Graphical and Computational Statistics}.
}
\keyword{Bounded score function, Generalized information criterion, Generalized linear model, Robust estimating equation, Robust quasi-likelihood, Smoothing parameter selection}
\examples{
# see example in function robustgam
}